refactor: move hook to contexts
